+++ src/context/UserInfoContext.tsx
@@ -1,0 +1,14 @@
+import { createContext } from "react";
+
+type User = {
+  email: string;
+};
+
+type UserInfoContextType = {
+  user: User | null;
+  setUser: React.Dispatch<React.SetStateAction<User | null>>;
+};
+
+export const UserInfoContext = createContext<UserInfoContextType | undefined>(
+  undefined
+);
